requirements.txt
boto3
flask
bcrypt
python-dotenv